Renamed package WIP. Added builder for AppSync functions
