+++ projects/code-from-asset-helper/.gitignore
