+++ projects/code-from-asset-helper/index.ts
@@ -1,0 +1,4 @@
+import { appSyncCodeFromAssetHelper, AssetHelperConfig, BuildMode, lambdaCodeFromAssetHelper } from "./lib/backend.utils";
+
+export { appSyncCodeFromAssetHelper, AssetHelperConfig, BuildMode, lambdaCodeFromAssetHelper };
+
