+++ projects/code-from-asset-helper/tsconfig.json
@@ -1,5 +1,8 @@
 {
   "compilerOptions": {
+    "types": [
+      "node"
+    ],
     "target": "ES2020",
     "module": "commonjs",
     "declaration": true,
