--- projects/lambda-code-from-asset-helper/index.ts
@@ -1,3 +1,0 @@
-import { AssetHelperConfig, BuildMode, lambdaCodeFromAssetHelper } from "./lib/backend.utils";
-
-export { AssetHelperConfig, BuildMode, lambdaCodeFromAssetHelper };
