--- projects/lambda-code-from-asset-helper/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@amplify-playground/lambda-code-from-asset-helper",
-  "version": "1.0.0",
-  "description": "A helper to build, bundle and copy a lambda function to the Amplify assets directory.",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com/naedx/amplify-playground.git"
-  },
-  "files": [
-    "dist/index.d.ts",
-    "dist/index.js",
-    "dist/lib/backend.utils.d.ts",
-    "dist/lib/backend.utils.js"
-  ],
-  "keywords": [
-    "Amplify",
-    "Lambda",
-    "esbuild",
-    "typescript"
-  ],
-  "author": "naedx",
-  "contributors": [
-    "MyNameIsTakenOMG"
-  ],
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/naedx/amplify-playground/issues"
-  },
-  "homepage": "https://github.com/naedx/amplify-playground/tree/dev/projects/lambda-code-from-asset-helper#readme",
-  "scripts": {
-    "prepublish": "tsc",
-    "build": "tsc"
-  },
-  "devDependencies": {
-    "@types/node": "^20.14.10",
-    "aws-cdk-lib": "^2.151.0",
-    "esbuild": "^0.23.0",
-    "typescript": "^5.5.4"
-  }
-}
